And fixed the lines where it was not tabbed properly.
Max, if you really insist of 'indent'ing everything, at least fix the
tool so that the source code can be seen whatever the tab size.

Ie something like that :

<TAB><TAB>sprintf(stderr, "Test %s",
<TAB><TAB>        variable);

Is VASTLY better than what indent does :
<TAB><TAB>sprintf(stderr, "Test %s",
<TAB><TAB><TAB><TAB><TAB>variable);

Wich is unreadable if you do not have the TAB size that YOU prefer
(and maybe not the rest of the Internet out there).

svn-id: r20