Detailed documentation on disassembler implementation
To compile to PDF, run "make doc", or go to the documentation directory
and run "make". If your system does not know the rm command, you may
want to change decompiler/doc/Makefile to use the equivalent for your
system to clean up all the temporary files from the LaTeX compilation.

svn-id: r49467