Allow engines to stop decompilation between steps
Engines can state they do not support code flow analysis or code
generation, and the decompiler will then stop before the unsupported
step.

This makes it easier to add and debug engines, as you don't have to
add the right command line parameter every time you run the program.

If the methods are not overridden, the default is to proceed.

svn-id: r51286