DECOMPILER: Add talkie flag to engine
Using a command-line switch (-t), you can now set a flag in an engine
that the script is from a talkie version of a game. This information is
needed for some opcodes in some engines, e.g. o2_npcChat in Kyra:HoF,
which takes an extra parameter in the talkie version.

svn-id: r51836