TOOLS: Fix extract_mohawk tool to allow extraction of resources with '/' or '\' in names.
The extraction of the resources with forward slashes in names was not possible as these are invalid in paths as they are directory seperators on Unix.

This commit fixes this by replacing them with backslashes, which can be present in filenames.

However in order to allow this operation to be unambiguously reversed by construct_mohawk, it is also necessary to pad any backslashes present prior to this to allow them to be disambiguated.

svn-id: r55592