DECOMPILER: Pure grouping option
The existing grouping algorithm doesn't work with non-stack-based
engines. Because dot is unable the large amount of vertices you get as a
result, this prevents visualization of the code flow. To get around
that, pure grouping provides an alternative which only considers
branches. This creates the minimal amount of vertices in the graph,
making it much easier for dot to layout the graph.

svn-id: r55832