TOOLS: Fix handling of input directory names
The code was supposed to add a '/' at the end of the input directories
provided by the user (if there wasn't one already) but it didn't due to
two different bugs. std::string::operator[] does not change the length
of the string, so it is not a good idea to overwrite '\0' with it. Also the
check to know if the path ended with a '/' was the wrong way out).