TOOLS: Improve behavior of Filename for directories
When creating a Filename with a path corresponding to a directory
but without a '/' or '\\' at the end the last part is treated as the file
name and not a directory and setFullName() does not behaves as
expected. Now when setting the path in Filename if it corresponds to
an existing directory it makes sure it ends with a '/' or '\\'.