TOOLS: Improve input file check in the GUI Tools
It now uses the same check that as been recently added when parsing
the input files in the CLI Tools. This is the first part of a fix for tools
expecting multiple inputs, as depending on the first file selected by
the user it could crash the application when trying to run the tool.