TOOLS: Improve tools detection from input file in GUI
The tools that expected a directory were never detected because the
user could only select a file and not a directory. Now he still needs
to select a file but the directory that contains it is also checked and
the file name is stripped if needed to run the tool.

An alternative would have been to let the user select a directory, but
that would have meant the user needed to know if the tool expect a file
or a directory as input. Also because there is no wxPicker to select either
a file or a directory this would have needed additional radio buttons to
show either a file picker or a directory picker. All in all always asking for
a file here seemed more user friendly.