TOOLS: Fix extra inputs page in GUI
It now matches the inputs already given with the expected inputs from
the tool to know which ones are missing. Also it now shows a picker
for all the inputs even those for which we already have an input file
(but they are initialized with the path we have). This prepares the next
commit.