TOOLS: GUI: Go directly to extra input page after tool selection
For tools expecting multiple inputs, after selecting the tool the user
had first to go to the normal input page to select one input and then
to the extra input page to select the other inputs. Now he gets directly
to the extra input page where he can select all the inputs.

The other input page is still used when the user decide to select a file
rather than a tool, the tool is then guessed from the file and the user
gets to the extra inputs page if the tools require other inputs (as before).