TOOLS: Fix crash when compiling with wxWidgets 2.9
The issue occurred on the Intro page when clicking on the Compress,
Extract or Advanced button. The page and therefore the buttons were
destroyed from the event handler of the button that was clicked. While
this work with wxWidgets 2.8 this seems dangerous. And this caused
a crash with wxWidgets 2.9 (at least on mac-cocoa). The page deletion
is now delayed.