TOOLS: Add MP3/OGG/FLAC signature to compressed Tinsel .smk files
Other than matching the signature in the .idx file, this ensures
that no sound begins at offset 0. This is good because offset 0
signifies that there is no sound. This caused the first line of
speech as Rincewind wakes up at the beginning of Discworld 1 to
only be writte, not spoken. This should not break backwards
compatibility in any way.