SWORD1: Improve support for sub-directory structure
When using a sub-directory structure with SPEECH and MUSIC
sub-directories, to compress both speech and music the user needed
to select either a cluster file or the swordres.rif file in the parent
directory. Now it should work if those files are in a Clusters
sub-directory and when selecting one of the speech file.