SWORD1: Fix memory corruption when trying to guess file endianness
The issue was that it was trying both little and big endian to check
which one seemed to be the most likely, but with the wrong
endianness it could write data beyond the end of the allocated
memory when decompressing the speech data as it was using an
incorrect size. This commits adds sanity checks to avoid this.