CONFIGURE: Ignore options valid for Autotools configure
RPM-based distributions come with a configure macro, that sets up the
whole build environment.  In order to be able to use this macro, some
options, which are valid for the configure script generated by GNU
Autotools, should not error-out when preparing the build stage.

Backported (and slightly modified) from scummvm/scummvm@826b7bfe7