TOOLS: Improve handling of inputs for tools that expect a directory as input
The main issue here is that the GUI Tools, if we use the default
wizard and do not first select a tool, has a file selector for
the input. It then tries to handle the case where the tool expect
a directory as input, but this was not properly handled and did
not work. The user had to manually stip the file name from the
input after selecting a file. It now works properly when giving
a file as input to a tool that expects a directory.