PRINCE: Improve input handling for the pack extract and pack tools
Those tools expect a directory as input, and where setting the
expect input pattern as such. However they were not setting the
file flag to false.

Also the inspectInput implementations now check files in the
given directory to be more discriminatory and not just accept
any directory as a perfect match.