TINSEL: fix FLAC package creation
A minor error in the code creates an .IDX file with corrupt data at the beginning of the file. This corrupt .IDX starts with `FLACFLAC` instead of `FLAC`. and the first proper index is blank.

Manually removing the erroneous `FLAC` from the .IDX file, allows the file to work in ScummVM, but some corruption has still occurred.

This can be most easily evidenced with the line "Rincewind, Rincewind, wake up!" in the intro sequence, which is missing given the above circumstances.

This is fixed with this commit.

Note: this also prepends `FLAC` to the .SMP file, to match the OGG and MP3 packages.